enhance login buttons handling
