+++ presentation/designSystem/src/main/kotlin/com/berlin/aflami/component/SpinningProgressIndicatorLines.kt
@@ -1,4 +1,4 @@
-package com.example.aflami.ui.component
+package com.berlin.aflami.component
 
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.animateFloat
