+++ presentation/designSystem/src/main/kotlin/com/berlin/aflami/component/buttons/FabButton.kt
@@ -1,31 +1,36 @@
-package com.example.aflami.ui.component.buttons
+package com.berlin.aflami.component.buttons
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.material3.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.aflami.designsystem.theme.Theme
+import com.berlin.aflami.ui.theme.Theme
 
 @Composable
-fun PrimaryButton(
+fun FabButton(
     onClick: () -> Unit,
     modifier: Modifier = Modifier,
     state: ButtonState = ButtonState.IDLE,
     enabled: Boolean = state != ButtonState.DISABLED,
     shape: Shape = ButtonDefaults.defaultShape,
-    contentPadding: PaddingValues = ButtonDefaults.defaultPadding,
+    contentPadding: PaddingValues = ButtonDefaults.defaultFabPadding,
     buttonColors: ButtonColors = ButtonDefaults.colors(),
     content: @Composable RowScope.() -> Unit
 ) {
+    val currentContent = if (state == ButtonState.LOADING) {{}}else content
+
     DefaultButton(
         onClick = onClick,
         modifier = modifier,
         state = state,
         enabled = enabled,
+        isFabType = true,
         contentPadding = contentPadding,
         colors = buttonColors.copy(
             backgroundGradient = Brush.horizontalGradient(
@@ -33,61 +38,59 @@ fun PrimaryButton(
             )
         ),
         shape = shape,
-        content = content
+        content = currentContent
     )
 }
 
-@Preview(showSystemUi = false, showBackground = true)
+@Preview
 @Composable
-private fun PrimaryButtonPreview() {
-    PrimaryButton(
+private fun FabButtonPreview() {
+    FabButton(
         onClick = {},
         content = {
-            Text(
-                text = "Button",
+            Icon(
+                imageVector = Icons.Default.Add,
+                contentDescription = null
             )
         }
     )
 }
 
-@Preview(showSystemUi = false, showBackground = true)
+@Preview
 @Composable
-private fun PrimaryButtonLoadingPreview() {
-    PrimaryButton(
+private fun FabButtonLoadingPreview() {
+    FabButton(
         onClick = {},
         state = ButtonState.LOADING,
-        content = {
-            Text(
-                text = "Button",
-            )
-        }
+        content = {}
     )
 }
 
-@Preview(showSystemUi = false, showBackground = true)
+@Preview
 @Composable
-private fun PrimaryButtonDisabledPreview() {
-    PrimaryButton(
+private fun FabButtonDisabledPreview() {
+    FabButton(
         onClick = {},
         state = ButtonState.DISABLED,
         content = {
-            Text(
-                text = "Button",
+            Icon(
+                imageVector = Icons.Default.Add,
+                contentDescription = null
             )
         }
     )
 }
 
-@Preview(showSystemUi = false, showBackground = true)
+@Preview
 @Composable
-private fun PrimaryButtonErrorPreview() {
-    PrimaryButton(
+private fun FabButtonErrorPreview() {
+    FabButton(
         onClick = {},
         state = ButtonState.ERROR,
         content = {
-            Text(
-                text = "Button",
+            Icon(
+                imageVector = Icons.Default.Add,
+                contentDescription = null
             )
-        }
-    )
+        })
 }
