+++ presentation/designSystem/src/main/kotlin/com/berlin/aflami/component/buttons/NegativeButton.kt
@@ -1,96 +1,91 @@
-package com.example.aflami.ui.component.buttons
+package com.berlin.aflami.component.buttons
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.aflami.designsystem.theme.Theme
+import com.berlin.aflami.ui.theme.Theme
 
 @Composable
-fun FabButton(
+fun NegativeButton(
     onClick: () -> Unit,
     modifier: Modifier = Modifier,
     state: ButtonState = ButtonState.IDLE,
     enabled: Boolean = state != ButtonState.DISABLED,
     shape: Shape = ButtonDefaults.defaultShape,
-    contentPadding: PaddingValues = ButtonDefaults.defaultFabPadding,
+    contentPadding: PaddingValues = ButtonDefaults.defaultPadding,
     buttonColors: ButtonColors = ButtonDefaults.colors(),
     content: @Composable RowScope.() -> Unit
 ) {
-    val currentContent = if (state == ButtonState.LOADING) {{}}else content
-
     DefaultButton(
         onClick = onClick,
         modifier = modifier,
         state = state,
         enabled = enabled,
-        isFabType = true,
         contentPadding = contentPadding,
         colors = buttonColors.copy(
-            backgroundGradient = Brush.horizontalGradient(
-                Theme.color.gradientColors.pointsOverly
-            )
+            backgroundColor = Theme.color.statusColors.redVariant,
+            contentColor =Theme.color.statusColors.redAccent,
         ),
         shape = shape,
-        content = currentContent
+        content = content
     )
 }
 
-@Preview
+@Preview(showSystemUi = false, showBackground = true)
 @Composable
-private fun FabButtonPreview() {
-    FabButton(
+private fun NegativeButtonPreview() {
+    NegativeButton(
         onClick = {},
         content = {
-            Icon(
-                imageVector = Icons.Default.Add,
-                contentDescription = null
+            Text(
+                text = "Button",
             )
         }
     )
 }
 
-@Preview
+@Preview(showSystemUi = false, showBackground = true)
 @Composable
-private fun FabButtonLoadingPreview() {
-    FabButton(
+private fun NegativeButtonLoadingPreview() {
+    NegativeButton(
         onClick = {},
         state = ButtonState.LOADING,
-        content = {}
+        content = {
+            Text(
+                text = "Button",
+            )
+        }
     )
 }
 
-@Preview
+@Preview(showSystemUi = false, showBackground = true)
 @Composable
-private fun FabButtonDisabledPreview() {
-    FabButton(
+private fun NegativeButtonDisabledPreview() {
+    NegativeButton(
         onClick = {},
         state = ButtonState.DISABLED,
         content = {
-            Icon(
-                imageVector = Icons.Default.Add,
-                contentDescription = null
+            Text(
+                text = "Button",
             )
         }
     )
 }
 
-@Preview
+@Preview(showSystemUi = false, showBackground = true)
 @Composable
-private fun FabButtonErrorPreview() {
-    FabButton(
+private fun NegativeButtonErrorPreview() {
+    NegativeButton(
         onClick = {},
         state = ButtonState.ERROR,
         content = {
-            Icon(
-                imageVector = Icons.Default.Add,
-                contentDescription = null
+            Text(
+                text = "Button",
             )
-        })
+        }
+    )
 }
