+++ presentation/designSystem/src/main/kotlin/com/berlin/aflami/component/buttons/SecondaryButton.kt
@@ -1,5 +1,6 @@
-package com.example.aflami.ui.component.buttons
+package com.berlin.aflami.component.buttons
 
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.material3.Text
@@ -8,15 +9,20 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.aflami.designsystem.theme.Theme
+import androidx.compose.ui.unit.dp
+import com.berlin.aflami.ui.theme.Theme
 
 @Composable
-fun TextButton(
+fun SecondaryButton(
     onClick: () -> Unit,
     modifier: Modifier = Modifier,
     state: ButtonState = ButtonState.IDLE,
     enabled: Boolean = state != ButtonState.DISABLED,
     shape: Shape = ButtonDefaults.defaultShape,
+    border: BorderStroke? = BorderStroke(
+        1.dp,
+        Theme.color.disable
+    ),
     contentPadding: PaddingValues = ButtonDefaults.defaultPadding,
     buttonColors: ButtonColors = ButtonDefaults.colors(),
     content: @Composable RowScope.() -> Unit
@@ -28,13 +34,14 @@ fun TextButton(
         enabled = enabled,
         contentPadding = contentPadding,
         colors = buttonColors.copy(
-            backgroundColor = Color.Transparent,
-            errorBackgroundColor = Color.Transparent,
-            disabledBackgroundColor = Color.Transparent,
+            backgroundColor = Theme.color.primaryVariant,
             contentColor = Theme.color.primary,
-            disabledContentColor =Theme.color.disable ,
+            disabledBackgroundColor = Color.Transparent,
+            disabledContentColor = Theme.color.stroke,
+            errorBackgroundColor = Color.Transparent,
             errorContentColor = Theme.color.statusColors.redAccent,
         ),
+        border = border,
         shape = shape,
         content = content
     )
@@ -42,8 +49,8 @@ fun TextButton(
 
 @Preview(showSystemUi = false, showBackground = true)
 @Composable
-private fun TextButtonPreview() {
-    TextButton(
+private fun SecondaryButtonPreview() {
+    SecondaryButton(
         onClick = {},
         content = {
             Text(
@@ -55,8 +62,8 @@ private fun TextButtonPreview() {
 
 @Preview(showSystemUi = false, showBackground = true)
 @Composable
-private fun TextButtonLoadingPreview() {
-    TextButton(
+private fun SecondaryButtonLoadingPreview() {
+    SecondaryButton(
         onClick = {},
         state = ButtonState.LOADING,
         content = {
@@ -67,10 +74,10 @@ private fun TextButtonLoadingPreview() {
     )
 }
 
-@Preview(showBackground = true)
+@Preview(showSystemUi = false, showBackground = true)
 @Composable
-private fun TextButtonDisabledPreview() {
-    TextButton(
+private fun SecondaryButtonDisabledPreview() {
+    SecondaryButton(
         onClick = {},
         state = ButtonState.DISABLED,
         content = {
@@ -81,10 +88,10 @@ private fun TextButtonDisabledPreview() {
     )
 }
 
-@Preview(showBackground = true, showSystemUi = false)
+@Preview(showSystemUi = false, showBackground = true)
 @Composable
-private fun TextButtonErrorPreview() {
-    TextButton(
+private fun SecondaryButtonErrorPreview() {
+    SecondaryButton(
         onClick = {},
         state = ButtonState.ERROR,
         content = {
