+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/login/LoginUiState.kt
@@ -2,6 +2,7 @@ package com.berlin.aflami.viewmodel.login
 
 data class LoginUiState(
     val formUiState: FormUiState = FormUiState(),
+    val isLoginButtonEnabled: Boolean = false,
     val isLoading: Boolean = false,
     val isError: Boolean = false,
 )
